refactor : 7주차 멘토링에 따른 랭킹 서비스 계층 이동
